% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_rand_index.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Rand Index (ARI)}
\usage{
adjusted_rand_index(input, label_true, reduction)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{label_true}{A string specifying cell types.}

\item{reduction}{A string specifying the name of dimensional reduction.}
}
\value{
A numeric value
}
\description{
The Adjusted Rand Index (ARI) is a metric used to measure the similarity
between the clustering and ground truth labels.
}
\details{
After computing clustering using Leiden algorithm, the ARI metric is computed
using the `adjusted_rand_score` function from the Python sklearn package.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
ari <- adjusted_rand_index(input = sim, label_true = "Group",
                           reduction = "PCA")

}
