% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_out.R
\docType{class}
\name{AltOutput-class}
\alias{AltOutput-class}
\alias{AltOutput}
\title{AltOutput class}
\description{
Class of an alternative output, that contains three slots: corrected matrix, embedding, and metadata.
}
\section{Slots}{

\describe{
\item{\code{corrected}}{ANY.}

\item{\code{embedding}}{matrix.}

\item{\code{meta}}{data.frame.}
}}

