% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_silhouette_width.R
\name{average_silhouette_width}
\alias{average_silhouette_width}
\title{Average Silhouette Width}
\usage{
average_silhouette_width(
  input,
  label_true,
  reduction,
  metric = "euclidean",
  sample_size = NULL,
  random_state = NULL
)
}
\arguments{
\item{input}{A `SingleCellExperiment` object.}

\item{label_true}{A string specifying cell types.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{metric}{The metric to use when calculating distance between instances.}

\item{sample_size}{The size of the sample to use when computing the Silhouette Coefficient on a random subset of the data.}

\item{random_state}{Determines random number generation for selecting a subset of samples.}
}
\description{
Average Silhouette Width
}
