% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{batch_params}
\alias{batch_params}
\title{Batch effects strength}
\usage{
batch_params(norm_counts, batch)
}
\arguments{
\item{norm_counts}{Normalized counts matrix.}

\item{batch}{A string specifying the batch variable.}
}
\value{
Median of symmetric Kullback-Leibler (KL) divergence
}
\description{
Batch effects strength
}
