% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/batchCorrect.R
\name{batchCorrect}
\alias{batchCorrect}
\alias{batchCorrect,LimmaParams-method}
\alias{batchCorrect,CombatParams-method}
\title{batchCorrect}
\usage{
batchCorrect(input, batch, params)

\S4method{batchCorrect}{LimmaParams}(input, batch, params)

\S4method{batchCorrect}{CombatParams}(input, batch, params)
}
\arguments{
\item{input}{A SingleCellExperiment, Seurat or AnnData objects can be supplied.}

\item{batch}{A string specifying the batch for each cell.}

\item{params}{A \linkS4class{BatChefParams} object specifying the batch correction method.}
}
\description{
Generic function for defining batch effect correction methods.

A common interface for single-cell batch correction methods.
}
