% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisi-package.R
\docType{package}
\name{BatChef}
\alias{BatChef-package}
\alias{BatChef}
\title{lisi}
\description{
Local neighborhood diversity for scRNAseq
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zuinelena3/BatChef}
  \item Report bugs at \url{https://github.com/zuinelena3/BatChef/issues}
}

}
\author{
\strong{Maintainer}: Elena Zuin \email{elena.zuin.3@phd.unipd.it} (\href{https://orcid.org/0009-0006-3060-4835}{ORCID}) (Department of Biology, University of Padova)

Other contributors:
\itemize{
  \item Chiara Romualdi (Department of Biology, University of Padova) [contributor]
  \item Davide Risso (Department of Statistical Sciences, University of Padova) [contributor]
  \item Gabriele Sales (Department of Biology, University of Padova) [contributor]
}

}
