% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknnInput.R
\name{bbknnInput}
\alias{bbknnInput}
\alias{bbknnInput,Seurat-method}
\alias{bbknnInput,Seurat,Seurat-method}
\alias{bbknnInput,SingleCellExperiment-method}
\alias{bbknnInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{bbknnInput,AnnDataR6-method}
\alias{bbknnInput,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
bbknnInput(input, batch, reduction)

\S4method{bbknnInput}{Seurat}(input, batch, reduction)

\S4method{bbknnInput}{SingleCellExperiment}(input, batch, reduction)

\S4method{bbknnInput}{AnnDataR6}(input, batch, reduction)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{reduction}{reduction}
}
\description{
Title
}
