% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknnRun.R
\name{bbknnRun}
\alias{bbknnRun}
\title{BBKNN method}
\usage{
bbknnRun(
  input,
  batch,
  neighbors_within_batch = 3,
  n_pcs = 50,
  trim = NULL,
  computation = "annoy",
  annoy_n_trees = 10,
  pynndescent_n_neighbors = 30,
  pynndescent_random_state = 0,
  metric = "euclidean",
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  approx = NULL,
  use_annoy = NULL,
  use_faiss = NULL,
  scanpy_logging = FALSE
)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{neighbors_within_batch}{neighbors_within_batch}

\item{n_pcs}{n_pcs}

\item{trim}{trim}

\item{computation}{computation}

\item{annoy_n_trees}{annoy_n_trees}

\item{pynndescent_n_neighbors}{pynndescent_n_neighbors}

\item{pynndescent_random_state}{pynndescent_random_state}

\item{metric}{metric}

\item{set_op_mix_ratio}{set_op_mix_ratio}

\item{local_connectivity}{local_connectivity}

\item{approx}{approx}

\item{use_annoy}{use_annoy}

\item{use_faiss}{use_faiss}

\item{scanpy_logging}{scanpy_logging}
}
\description{
BBKNN method
}
