% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_params.R
\name{capture_params}
\alias{capture_params}
\title{Capture parameters of the correction methods}
\usage{
capture_params(target, params)
}
\arguments{
\item{target}{Target parameters}

\item{params}{Parameters}
}
\description{
Capture parameters of the correction methods
}
