% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering after integration reduction}
\usage{
clustering(input, label_true, reduction)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{label_true}{A string specifying cell types.}

\item{reduction}{A string specifying the dimensional reduction.}
}
\description{
Compute Leiden clustering. The resolution was chosen using the NMI metric.
}
