% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustInput.R
\name{clustInput}
\alias{clustInput}
\alias{clustInput,Seurat-method}
\alias{clustInput,Seurat,Seurat-method}
\alias{clustInput,SingleCellExperiment-method}
\alias{clustInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{clustInput,AnnDataR6-method}
\alias{clustInput,AnnDataR6,AnnDataR6-method}
\title{Convert into a SingleCellExperiment object for clustering}
\usage{
clustInput(input, reduction)

\S4method{clustInput}{Seurat}(input, reduction)

\S4method{clustInput}{SingleCellExperiment}(input, reduction)

\S4method{clustInput}{AnnDataR6}(input, reduction)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{reduction}{A string specifying the batch for each cell.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object.
}
\description{
Convert into a SingleCellExperiment object for clustering
}
