% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatRun.R
\name{combatRun}
\alias{combatRun}
\title{ComBat method}
\usage{
combatRun(input, batch, assay_type = "counts", ...)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch for each cell.}

\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\value{
A corrected matrix
}
\description{
ComBat allows users to adjust for batch effects in datasets using an
empirical Bayes framework.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = FALSE,
                     output_format = "SingleCellExperiment")
combat <- combatRun(input = sim, batch = "Batch")

}
