% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lisi.R
\name{compute_lisi}
\alias{compute_lisi}
\title{Use this function to compute LISI scores of one or more labels.}
\usage{
compute_lisi(X, meta_data, label_colnames, perplexity = 30, nn_eps = 0)
}
\arguments{
\item{X}{A matrix with cells (rows) and features (columns).}

\item{meta_data}{A data frame with one row per cell.}

\item{label_colnames}{Which variables to compute LISI for.}

\item{perplexity}{The effective number of each cell's neighbors.}

\item{nn_eps}{Error bound for nearest neighbor search with \code{RANN:nn2()}.
Default of 0.0 implies exact nearest neighbor search.}
}
\value{
A data frame of LISI values. Each row is a cell and each
column is a different label variable.
}
\description{
Use this function to compute LISI scores of one or more labels.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(50, 50),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
red <- SingleCellExperiment::reducedDim(sim, "PCA")
lisi <- compute_lisi(X = red, meta_data = SingleCellExperiment::colData(sim),
                    label_colnames = "Group")

}
