% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Extract data characteristics}
\usage{
extract_features(input, batch)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object can be supplied.}

\item{batch}{A string specifying the batch variable.}
}
\value{
A data.frame that contains the features of input data.
}
\description{
Extract data characteristics
}
