% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fake_data}
\alias{fake_data}
\title{Simulated data}
\usage{
fake_data(nGenes, batchCells, group.prob)
}
\arguments{
\item{nGenes}{number of genes}

\item{batchCells}{number of cells for each batch}

\item{group.prob}{percent of cell types}
}
\description{
Simulated data
}
