% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMNNPost.R
\name{fastMNNPost}
\alias{fastMNNPost}
\alias{fastMNNPost,Seurat-method}
\alias{fastMNNPost,Seurat,Seurat-method}
\alias{fastMNNPost,SingleCellExperiment-method}
\alias{fastMNNPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{fastMNNPost,AnnDataR6-method}
\alias{fastMNNPost,AnnDataR6,AnnDataR6-method}
\title{Convert fastMNN output}
\usage{
fastMNNPost(input, output, method)

\S4method{fastMNNPost}{Seurat}(input, output, method)

\S4method{fastMNNPost}{SingleCellExperiment}(input, output, method)

\S4method{fastMNNPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{output}{fastMNN output: a \link[SingleCellExperiment]{SingleCellExperiment}
containing the reconstructed expression matrix and the corrected dimensionality
reduction space.}

\item{method}{A string specifying the correction method.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
\description{
Convert fastMNN output into a \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object
}
