% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMNNRun.R
\name{fastMNNRun}
\alias{fastMNNRun}
\title{fastMNN method}
\usage{
fastMNNRun(input, batch, ...)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{...}{Named arguments to pass to individual methods upon dispatch}
}
\description{
fastMNN method
}
