% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{groups_params}
\alias{groups_params}
\title{Groups parameters}
\usage{
groups_params(input, normalized)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{normalized}{A normalized counts matrix.}
}
\value{
The number of clusters.
}
\description{
Groups parameters
}
