% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonyRun.R
\name{harmonyRun}
\alias{harmonyRun}
\title{Harmony method}
\usage{
harmonyRun(input, batch, ...)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{...}{Named arguments to pass to individual methods upon dispatch}
}
\description{
Harmony method
}
