% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{kl_gamma}
\alias{kl_gamma}
\title{Kullback-Leibler (KL) divergence) between two Gamma distributions}
\usage{
kl_gamma(p, q)
}
\arguments{
\item{p}{Gamma parameters (shape and rate) of probability distribution P}

\item{q}{Gamma parameters (shape and rate) of probability distribution Q}
}
\value{
Kullback-Leibler (KL) divergence) between two Gamma distributions
}
\description{
Kullback-Leibler (KL) divergence) between two Gamma distributions
}
