% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{knn_index_dist}
\alias{knn_index_dist}
\title{Retrive knn index and distances from BBKNN output}
\usage{
knn_index_dist(dist)
}
\arguments{
\item{dist}{A dgRMatrix of pairwise distances}
}
\description{
Retrive knn index and distances from BBKNN output
}
