% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_clustering.R
\name{leiden_clustering}
\alias{leiden_clustering}
\title{Leiden clustering}
\usage{
leiden_clustering(
  input,
  label_true = NULL,
  reduction,
  nmi_compute = TRUE,
  resolution = NULL,
  k = 10,
  store = FALSE
)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{label_true}{A string specifying the ground truth label.}

\item{reduction}{A string specifying the dimensional reduction
on which the clustering analysis will be performed.}

\item{nmi_compute}{A Boolean value indicating NMI metric calculation to
identify the optimal clustering is to be performed (Default: TRUE).}

\item{resolution}{A numeric value specifying the resolution parameter.}

\item{k}{An integer scalar specifying the number of nearest neighbors.}

\item{store}{A Boolean value indicating whether cluster labels are stored
within the input object (Default: FALSE).}
}
\value{
A A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object that contains the cluster labels.
}
\description{
Leiden clustering algorithm.
}
\details{
The clustering algorithm can be executed by specifying either a single
resolution parameter or range of resolution parameters (from 0.1 to 2).
In the case of multiple resolutions, the clustering outcome that corresponds
to the highest Normalized Mutual Information (NMI) score is selected. Finding
the optimal clustering can be useful to compute the performance evaluation of
batch correction methods.
}
\examples{

sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
clust <- leiden_clustering(input = sim, reduction = "PCA",
                           nmi_compute = FALSE, resolution = 1)

}
