% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{lib_size_params}
\alias{lib_size_params}
\title{Library size parameters}
\usage{
lib_size_params(counts)
}
\arguments{
\item{counts}{Raw counts matrix.}
}
\value{
Library size parameters.
}
\description{
Library size parameters
}
