% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerInput.R
\name{ligerInput}
\alias{ligerInput}
\alias{ligerInput,Seurat-method}
\alias{ligerInput,Seurat,Seurat-method}
\alias{ligerInput,SingleCellExperiment-method}
\alias{ligerInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{ligerInput,AnnDataR6-method}
\alias{ligerInput,AnnDataR6,AnnDataR6-method}
\title{Convert to a LIGER compatible object}
\usage{
ligerInput(input, batch, features, ...)

\S4method{ligerInput}{Seurat}(input, batch, features, ...)

\S4method{ligerInput}{SingleCellExperiment}(input, batch, features, ...)

\S4method{ligerInput}{AnnDataR6}(input, batch, features, ...)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{batch}{A string specifying the batch variable.}

\item{features}{Vector of features to use.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
Convert to a LIGER compatible object
}
