% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerPost.R
\name{ligerPost}
\alias{ligerPost}
\alias{ligerPost,Seurat-method}
\alias{ligerPost,Seurat,Seurat-method}
\alias{ligerPost,SingleCellExperiment-method}
\alias{ligerPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{ligerPost,AnnDataR6-method}
\alias{ligerPost,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
ligerPost(input, output, method)

\S4method{ligerPost}{Seurat}(input, output, method)

\S4method{ligerPost}{SingleCellExperiment}(input, output, method)

\S4method{ligerPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}
}
\description{
Title
}
