% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaRun.R
\name{limmaRun}
\alias{limmaRun}
\title{limma method}
\usage{
limmaRun(input, batch, assay_type = "logcounts", ...)
}
\arguments{
\item{input}{A SingleCellExperiment object.}

\item{batch}{A string specifying the batch for each cell.}

\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
limma method
}
