% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearInput.R
\name{linearInput}
\alias{linearInput}
\alias{linearInput,Seurat-method}
\alias{linearInput,Seurat,Seurat-method}
\alias{linearInput,SingleCellExperiment-method}
\alias{linearInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{linearInput,AnnDataR6-method}
\alias{linearInput,AnnDataR6,AnnDataR6-method}
\title{Convert into a SingleCellExperiment object for linear model based methods}
\usage{
linearInput(input, batch)

\S4method{linearInput}{Seurat}(input, batch)

\S4method{linearInput}{SingleCellExperiment}(input, batch)

\S4method{linearInput}{AnnDataR6}(input, batch)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{batch}{A string specifying the batch for each cell.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object.
}
\description{
Convert into a SingleCellExperiment object for linear model based methods
}
