% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearPost.R
\name{linearPost}
\alias{linearPost}
\alias{linearPost,Seurat-method}
\alias{linearPost,Seurat,Seurat-method}
\alias{linearPost,SingleCellExperiment-method}
\alias{linearPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{linearPost,AnnDataR6-method}
\alias{linearPost,AnnDataR6,AnnDataR6-method}
\title{Convert the output of linear models}
\usage{
linearPost(input, output, method)

\S4method{linearPost}{Seurat}(input, output, method)

\S4method{linearPost}{SingleCellExperiment}(input, output, method)

\S4method{linearPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{output}{Linear models output}

\item{method}{A string specifying the correction method.}
}
\description{
Convert the output of linear models (limma and ComBat) into a SingleCellExperiment, Seurat or Anndata objects.
}
