% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_inverse_simpson_index.R
\name{local_inverse_simpson_index}
\alias{local_inverse_simpson_index}
\title{Local Inverse Simposon Index}
\usage{
local_inverse_simpson_index(
  input,
  label_true,
  reduction,
  meta_data = colData(input),
  perplexity = 30,
  nn_eps = 0
)
}
\arguments{
\item{input}{input}

\item{label_true}{label_true}

\item{reduction}{reduction}

\item{meta_data}{meta_data}

\item{perplexity}{perplexity}

\item{nn_eps}{nn_eps}
}
\description{
Local Inverse Simposon Index
}
