% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predition_plot.R
\name{log_transf}
\alias{log_transf}
\title{A shifted log transformation}
\usage{
log_transf(mat, base = exp(1))
}
\arguments{
\item{mat}{A matrix of data characteristics}

\item{base}{logarithm base}
}
\value{
A shifted log transformation matrix.
}
\description{
A shifted log transformation
}
