% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{mean_params}
\alias{mean_params}
\title{Mean parameters}
\usage{
mean_params(normalized)
}
\arguments{
\item{normalized}{Normalized counts matrix.}
}
\value{
Mean rate parameter.
}
\description{
Mean parameters
}
