% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_params.R
\name{merge_params}
\alias{merge_params}
\title{Parameters merging}
\usage{
merge_params(base, extra, class)
}
\arguments{
\item{base}{base params}

\item{extra}{extra params}

\item{class}{class}
}
\description{
Parameters merging
}
