% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodParam.R, R/classes.R
\docType{class}
\name{limmaMethod}
\alias{limmaMethod}
\alias{combatMethod}
\alias{seuratv3Method}
\alias{seuratv5Method}
\alias{fastMNNMethod}
\alias{harmonyMethod}
\alias{scanoramaMethod}
\alias{bbknnMethod}
\alias{scVIMethod}
\alias{scMergeMethod}
\alias{MethodParam-class}
\alias{limmaMethod-class}
\alias{combatMethod-class}
\alias{seuratv3Method-class}
\alias{seuratv5Method-class}
\alias{fastMNNMethod-class}
\alias{harmonyMethod-class}
\alias{scanoramaMethod-class}
\alias{bbknnMethod-class}
\alias{scVIMethod-class}
\alias{scMergeMethod-class}
\title{MethodParam methods}
\usage{
limmaMethod(...)

combatMethod(...)

seuratv3Method(...)

seuratv5Method(...)

fastMNNMethod(...)

harmonyMethod(...)

scanoramaMethod(...)

bbknnMethod(...)

scVIMethod(...)

scMergeMethod(...)
}
\arguments{
\item{...}{Named arguments to pass to individual methods upon dispatch}
}
\description{
Constructors and methods for the method parameter classes.
}
