% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Performance evaluation metrics.}
\usage{
metrics(
  input,
  batch,
  group,
  reduction,
  rep = 10,
  mc_cores = 1,
  nmi_compute = TRUE,
  resolution = NULL,
  average_method = "arithmetic",
  metric = "euclidean",
  sample_size = NULL,
  random_state = NULL,
  meta_data = colData(input),
  perplexity = 30,
  nn_eps = 0
)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{batch}{A string specifying batch variable.}

\item{group}{A string specifying the ground truth label.}

\item{reduction}{A string specifying the dimensional reduction.
on which the clustering analysis will be performed.}

\item{rep}{Number of times the Wasserstein distance is calculated.}

\item{mc_cores}{The number of cores to use.}

\item{nmi_compute}{A Boolean value indicating NMI metric calculation to
identify the optimal clustering is to be performed (Default: TRUE).}

\item{resolution}{A numeric value specifying the resolution parameter.}

\item{average_method}{How to compute the normalizer in the denominator.}

\item{metric}{The metric to use when calculating distance between instances.}

\item{sample_size}{The size of the sample to use when computing the
Silhouette Coefficient on a random subset of the data.}

\item{random_state}{Determines random number generation for selecting
a subset of samples.}

\item{meta_data}{A data frame with one row per cell.}

\item{perplexity}{The effective number of each cell's neighbors.}

\item{nn_eps}{Error bound for nearest neighbor search with `RANN:nn2()`.}
}
\value{
A data.frame object
}
\description{
Performance evaluation metrics: Wasserstein distance, Local Inverse Simpson's
Index, Average Silhouette Width, Adjusted Rand Index, and
Normalized Mutual Information.
}
\details{
This function performs Leiden clustering to facilitate the computation of
the Adjusted Rand Index (ARI) and Normalized Mutual Information (NMI) metrics.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 110),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
metrics <- metrics(input = sim, batch = "Batch", group = "Group",
                   reduction = "PCA", rep = 5)

}
