% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_mutual_info.R
\name{normalized_mutual_info}
\alias{normalized_mutual_info}
\title{Normalized Mutual Information}
\usage{
normalized_mutual_info(
  input,
  label_true,
  average_method = "arithmetic",
  reduction
)
}
\arguments{
\item{input}{input}

\item{label_true}{label_true}

\item{average_method}{average_method}

\item{reduction}{reduction}
}
\description{
Normalized Mutual Information
}
