% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{outlier_params}
\alias{outlier_params}
\title{Highly expressed genes probability}
\usage{
outlier_params(norm_counts)
}
\arguments{
\item{norm_counts}{A normalized matrix}
}
\value{
Highly expressed genes probability
}
\description{
Highly expressed genes probability
}
