% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_btc_factors}
\alias{params_btc_factors}
\title{Gamma parameters estimation of batch factors}
\usage{
params_btc_factors(factors)
}
\arguments{
\item{factors}{Batch factors}
}
\value{
a data.frame with estimated shape and rate parameters
}
\description{
Gamma parameters estimation of batch factors
}
