% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plot.R
\name{prediction_plot}
\alias{prediction_plot}
\title{Prediction plot}
\usage{
prediction_plot(params)
}
\arguments{
\item{params}{A data.frame of data characteristics}
}
\value{
A ggplot object
}
\description{
Prediction plot
}
