% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predition_plot.R
\name{predition_plot}
\alias{predition_plot}
\title{Prediction plot}
\usage{
predition_plot(params)
}
\arguments{
\item{params}{A data.frame of data characteristics}
}
\value{
A ggplot object
}
\description{
Prediction plot
}
