% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaInput.R
\name{scanoramaInput}
\alias{scanoramaInput}
\alias{scanoramaInput,Seurat-method}
\alias{anndatasInput,Seurat,Seurat-method}
\alias{scanoramaInput,SingleCellExperiment-method}
\alias{anndatasInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{scanoramaInput,AnnDataR6-method}
\alias{anndatasInput,AnnDataR6,AnnDataR6-method}
\title{Convert into a list of matrices}
\usage{
scanoramaInput(input, batch, assay.type = NULL)

\S4method{scanoramaInput}{Seurat}(input, batch, assay.type = NULL)

\S4method{scanoramaInput}{SingleCellExperiment}(input, batch, assay.type = NULL)

\S4method{scanoramaInput}{AnnDataR6}(input, batch, assay.type = NULL)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{batch}{A string specifying the batch for each cell.}

\item{assay.type}{layers name}
}
\description{
Convert into a list of matrices
}
