% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaRun.R
\name{scanoramaRun}
\alias{scanoramaRun}
\title{Scanorama method}
\usage{
scanoramaRun(
  input,
  return_dimred = FALSE,
  batch_size = 5000,
  verbose = TRUE,
  ds_names = NULL,
  dimred = 100,
  approx = TRUE,
  sigma = 15,
  alpha = 0.1,
  knn = 20,
  return_dense = FALSE,
  hvg = NULL,
  union = FALSE,
  seed = 0
)
}
\arguments{
\item{input}{List of expression matrices for each bacthes and the genes.}

\item{return_dimred}{A logical to returning integrated low-dimesional embeddings.}

\item{batch_size}{The batch size used in the alignment vector computation.}

\item{verbose}{Print progress bars and output.}

\item{ds_names}{Reports data set names in logging output}

\item{dimred}{Number of dimensions of corrected embedding.}

\item{approx}{A logical to use approximate nearest neighbors}

\item{sigma}{Correction smoothing parameter on Gaussian kernel.}

\item{alpha}{Alignment score minimum cutoff.}

\item{knn}{Number of nearest neighbors to use for matching.}

\item{return_dense}{A logical to return dense matrices.}

\item{hvg}{Use this number of top highly variable genes based on dispersion.}

\item{union}{A logical to consider the union of the genes. In alternative, the intersection is considered.}

\item{seed}{Random seed to use.}
}
\description{
Scanorama method
}
