% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceInput.R
\name{sceInput}
\alias{sceInput}
\alias{sceInput,Seurat-method}
\alias{sceInput,Seurat,Seurat-method}
\alias{sceInput,SingleCellExperiment-method}
\alias{sceInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{sceInput,AnnDataR6-method}
\alias{sceInput,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
sceInput(input, batch)

\S4method{sceInput}{Seurat}(input, batch)

\S4method{sceInput}{SingleCellExperiment}(input, batch)

\S4method{sceInput}{AnnDataR6}(input, batch)
}
\arguments{
\item{input}{input}

\item{batch}{batch}
}
\description{
Title
}
