% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMerge2Post.R
\name{scMerge2Post}
\alias{scMerge2Post}
\alias{scMerge2Post,Seurat-method}
\alias{scMerge2Post,Seurat,Seurat-method}
\alias{scMerge2Post,SingleCellExperiment-method}
\alias{scMerge2Post,SingleCellExperiment,SingleCellExperiment-method}
\alias{scMerge2Post,AnnDataR6-method}
\alias{scMerge2Post,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
scMerge2Post(input, output, method)

\S4method{scMerge2Post}{Seurat}(input, output, method)

\S4method{scMerge2Post}{SingleCellExperiment}(input, output, method)

\S4method{scMerge2Post}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}
}
\description{
Title
}
