% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMerge2Run.R
\name{scMerge2Run}
\alias{scMerge2Run}
\title{scMerge2 method}
\usage{
scMerge2Run(input, batch, assay_type = "logcounts", ...)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch variable.}

\item{assay_type}{A string specifying the assay to use for correction.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\value{
A list that contains the corrected gene expression matrix
}
\description{
scMerge2 is based on pseudo-replicates to remove unwanted batch effects.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(250, 200),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
scmerge2 <- scMerge2Run(input = sim, batch = "Batch", assay_type = "logcounts")

}
