% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratv3Input.R
\name{seuratv3Input}
\alias{seuratv3Input}
\alias{seuratv3Input,Seurat-method}
\alias{seuratv3Input,Seurat,Seurat-method}
\alias{seuratv3Input,SingleCellExperiment-method}
\alias{seuratv3Input,SingleCellExperiment,SingleCellExperiment-method}
\alias{seuratv3Input,AnnDataR6-method}
\alias{seuratv3Input,AnnDataR6,AnnDataR6-method}
\alias{seuratv3Input,list-method}
\alias{seuratv3Input,list,list-method}
\title{Convert into a Seurat V3 object}
\usage{
seuratv3Input(input, batch, features, pca_name = NULL)

\S4method{seuratv3Input}{Seurat}(input, batch, features, pca_name = NULL)

\S4method{seuratv3Input}{SingleCellExperiment}(input, batch, features, pca_name = NULL)

\S4method{seuratv3Input}{AnnDataR6}(input, batch, features, pca_name = NULL)

\S4method{seuratv3Input}{list}(input, batch, features, pca_name = NULL)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{batch}{A string specifying the batch variable.}

\item{features}{Vector of features to use.}

\item{pca_name}{A string specifying the PCA.}
}
\description{
Convert into a Seurat V3 object
}
