% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratV5Post.R
\name{seuratv5Post}
\alias{seuratv5Post}
\alias{seuratv5Post,Seurat-method}
\alias{seuratv5Post,Seurat,Seurat-method}
\alias{seuratv5Post,SingleCellExperiment-method}
\alias{seuratv5Post,SingleCellExperiment,SingleCellExperiment-method}
\alias{seuratv5Post,AnnDataR6-method}
\alias{seuratv5Post,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
seuratv5Post(input, output, method, name)

\S4method{seuratv5Post}{Seurat}(input, output, method, name)

\S4method{seuratv5Post}{SingleCellExperiment}(input, output, method, name)

\S4method{seuratv5Post}{AnnDataR6}(input, output, method, name)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}

\item{name}{reduction}
}
\description{
Title
}
