% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Function to simulate data}
\usage{
simulate_data(
  n_genes = 10000,
  batch_cells = 100,
  batch_fac_loc = 0.1,
  batch_fac_scale = 0.1,
  batch_rm_effect = FALSE,
  mean_rate = 0.3,
  mean_shape = 0.6,
  lib_loc = 11,
  lib_scale = 0.2,
  lib_norm = FALSE,
  out_prob = 0.05,
  out_fac_loc = 4,
  out_fac_scale = 0.5,
  group_prob = 1,
  de_prob = 0.1,
  de_down_prob = 0.5,
  de_fac_loc = 0.1,
  de_fac_scale = 0.4,
  bcv_common = 0.1,
  bcv_df = 60,
  dropout_type = "none",
  dropout_mid = 0,
  dropout_shape = -1,
  path_from = 0,
  path_n_steps = 100,
  path_skew = 0.5,
  path_nonlinear_prob = 0.1,
  path_sigma_fac = 0.8,
  compute_hvgs = FALSE,
  n_hvgs = 1000,
  num_threads = 1,
  compute_pca = FALSE,
  pca_ncomp = 10,
  output_format = "SingleCellExperiment",
  seed = 333
)
}
\arguments{
\item{n_genes}{Number of genes.}

\item{batch_cells}{Number of cells per batch.}

\item{batch_fac_loc}{Batch factor location parameter.}

\item{batch_fac_scale}{Batch factor scale parameter.}

\item{batch_rm_effect}{Remove batch effect.}

\item{mean_rate}{Mean rate.}

\item{mean_shape}{Mean shape.}

\item{lib_loc}{Library size location parameter.}

\item{lib_scale}{Library size scale parameter.}

\item{lib_norm}{Library size distribution.}

\item{out_prob}{Expression outlier probability.}

\item{out_fac_loc}{Expression outlier factor location.}

\item{out_fac_scale}{Expression outlier factor scale.}

\item{group_prob}{Group probabilities.}

\item{de_prob}{Differential expression probability.}

\item{de_down_prob}{Down-regulation probability.}

\item{de_fac_loc}{DE factor location.}

\item{de_fac_scale}{DE factor scale.}

\item{bcv_common}{Common biological coefficient of variation.}

\item{bcv_df}{BCV degrees of freedom.}

\item{dropout_type}{Dropout type.}

\item{dropout_mid}{Dropout mid point.}

\item{dropout_shape}{Dropout shape.}

\item{path_from}{Path origin.}

\item{path_n_steps}{Number of steps.}

\item{path_skew}{Path skew.}

\item{path_nonlinear_prob}{Non-linear probability.}

\item{path_sigma_fac}{Path skew.}

\item{compute_hvgs}{Boolean value. If TRUE, highly variable genes
will be selected Default is FALSE.}

\item{n_hvgs}{Number of highly variable genes.}

\item{num_threads}{Integer scalar specifying the number of threads to use.}

\item{compute_pca}{Boolean value. If TRUE, Principal Component Analysis (PCA)
will be computed. Default is FALSE.}

\item{pca_ncomp}{Number of principal component.}

\item{output_format}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.}

\item{seed}{Random seed.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
\description{
The function allows to simulated single-cell RNA-seq data using Splatter
package, normalize data, select highly variable genes and
compute Principal Component Analysis.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 1000,
                     compute_pca = FALSE)

}
