% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{simulated_data}
\alias{simulated_data}
\title{Function to simulate data}
\usage{
simulated_data(
  nGenes = 10000,
  batchCells = 100,
  batch.facLoc = 0.1,
  batch.facScale = 0.1,
  batch.rmEffect = FALSE,
  mean.rate = 0.3,
  mean.shape = 0.6,
  lib.loc = 11,
  lib.scale = 0.2,
  lib.norm = FALSE,
  out.prob = 0.05,
  out.facLoc = 4,
  out.facScale = 0.5,
  group.prob = 1,
  de.prob = 0.1,
  de.downProb = 0.5,
  de.facLoc = 0.1,
  de.facScale = 0.4,
  bcv.common = 0.1,
  bcv.df = 60,
  dropout.type = "none",
  dropout.mid = 0,
  dropout.shape = -1,
  path.from = 0,
  path.nSteps = 100,
  path.skew = 0.5,
  path.nonlinearProb = 0.1,
  path.sigmaFac = 0.8,
  seed = 333,
  ncomp = 10,
  n_hvgs = 500
)
}
\arguments{
\item{nGenes}{nGenes}

\item{batchCells}{batchCells}

\item{batch.facLoc}{batch.facLoc}

\item{batch.facScale}{batch.facScale}

\item{batch.rmEffect}{batch.rmEffect}

\item{mean.rate}{mean.rate}

\item{mean.shape}{mean.shape}

\item{lib.loc}{lib.loc}

\item{lib.scale}{lib.scale}

\item{lib.norm}{lib.norm}

\item{out.prob}{out.prob}

\item{out.facLoc}{out.facLoc}

\item{out.facScale}{out.facScale}

\item{group.prob}{group.prob}

\item{de.prob}{de.prob}

\item{de.downProb}{de.downProb}

\item{de.facLoc}{de.facLoc}

\item{de.facScale}{de.facScale}

\item{bcv.common}{bcv.common}

\item{bcv.df}{bcv.df}

\item{dropout.type}{dropout.type}

\item{dropout.mid}{dropout.mid}

\item{dropout.shape}{dropout.shape}

\item{path.from}{path.from}

\item{path.nSteps}{path.nSteps}

\item{path.skew}{path.skew}

\item{path.nonlinearProb}{path.nonlinearProb}

\item{path.sigmaFac}{path.sigmaFac}

\item{seed}{seed}

\item{ncomp}{ncomp}

\item{n_hvgs}{n_hvgs}
}
\description{
Function to simulate data
}
