% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested_method.R
\name{suggested_method}
\alias{suggested_method}
\title{Suggested method prediction}
\usage{
suggested_method(input, batch)
}
\arguments{
\item{batch}{A string specifying the batch variable.}

\item{A}{\link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}
}
\value{
A string specifying the suggested method.
}
\description{
Suggested method prediction
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 1000,
                     compute_pca = FALSE, output_format = "SingleCellExperiment")
pred <- suggested_method(input = sim, batch = "Batch")

}
