% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein_distance}
\alias{wasserstein_distance}
\title{Wasserstein distance}
\usage{
wasserstein_distance(input, batch, reduction, rep, mc.cores)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{reduction}{reduction}

\item{rep}{rep}

\item{mc.cores}{mc.cores}
}
\description{
Wasserstein distance
}
