% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{wasserstein_metric}
\alias{wasserstein_metric}
\title{Wasserstein distance}
\usage{
wasserstein_metric(obj, batch, reduction, rep)
}
\arguments{
\item{obj}{a SingleCellExperiment object}

\item{batch}{a string specifying the batch}

\item{reduction}{a string specifying the dimension reduction name}

\item{rep}{number of random selection replicates}
}
\description{
Wasserstein distance
}
