% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{winsorization}
\alias{winsorization}
\title{Winsorization}
\usage{
winsorization(norm_counts)
}
\arguments{
\item{norm_counts}{A normalized matrix}
}
\value{
Winsorized numeric vector
}
\description{
It's a transformation by limiting extreme values in data
to reduce the effect of outliers.
}
