% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_rand_index.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Rand Index (ARI)}
\usage{
adjusted_rand_index(input, label_true, reduction)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{label_true}{A string specifying cell types.}

\item{reduction}{A string specifying the dimensional reduction.}
}
\description{
Before ARI, the Leiden clustering is computed.
}
