% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_rand_index.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Rand Index (ARI)}
\usage{
adjusted_rand_index(
  input,
  label_true,
  reduction,
  resolution,
  k = 15,
  n_iter = -1
)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{label_true}{A string specifying the ground truth label.}

\item{reduction}{A string specifying the dimensional reduction
on which the clustering analysis will be performed.}

\item{resolution}{A numeric value specifying the resolution parameter.}

\item{k}{An integer scalar specifying the number of nearest neighbors.}

\item{n_iter}{Number of iterations of the Leiden clustering algorithm to perform.
Positive values above 2 define the total number of iterations to perform,
-1 has the algorithm run until it reaches its optimal clustering}
}
\value{
A numeric value
}
\description{
The Adjusted Rand Index (ARI) is a metric used to measure the similarity
between the clustering and ground truth labels.
}
\details{
After computing Leiden clustering algorithm, the ARI metric is computed.
}
\examples{
sim <- simulate_data(n_genes = 500, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
ari <- adjusted_rand_index(input = sim, label_true = "Group", reduction = "PCA",
                           resolution = 0.5, n_iter = 2)

}
