% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_rand_index.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Range Index}
\usage{
adjusted_rand_index(input, label_true, reduction)
}
\arguments{
\item{input}{input}

\item{label_true}{label_true}

\item{reduction}{reduction}
}
\description{
Adjusted Range Index
}
