% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndataInput.R
\name{anndataInput}
\alias{anndataInput}
\alias{anndataInput,Seurat-method}
\alias{anndataInput,Seurat,Seurat-method}
\alias{anndataInput,SingleCellExperiment-method}
\alias{anndataInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{anndataInput,AnnDataR6-method}
\alias{anndataInput,anndata.AnnData,anndata.AnnData-method}
\title{Title}
\usage{
anndataInput(input)

\S4method{anndataInput}{Seurat}(input)

\S4method{anndataInput}{SingleCellExperiment}(input)

\S4method{anndataInput}{AnnDataR6}(input)
}
\arguments{
\item{input}{input}
}
\description{
Title
}
