% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_silhouette_width.R
\name{average_silhouette_width}
\alias{average_silhouette_width}
\title{Average Silhouette Width}
\usage{
average_silhouette_width(
  input,
  label_true,
  reduction,
  metric = "euclidean",
  sample_size = NULL,
  random_state = NULL
)
}
\arguments{
\item{input}{input}

\item{label_true}{label_true}

\item{reduction}{reduction}

\item{metric}{metric}

\item{sample_size}{sample_size}

\item{random_state}{random_state}
}
\description{
Average Silhouette Width
}
