% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_silhouette_width.R
\name{average_silhouette_width}
\alias{average_silhouette_width}
\title{Average Silhouette Width (ASW)}
\usage{
average_silhouette_width(input, label_true, reduction, metric = "euclidean")
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{label_true}{A string specifying the ground truth label.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{metric}{The metric to use when calculating distance between instances.}
}
\value{
A numeric value
}
\description{
The average silhouette width for each cell is defined as the difference
between the average of  within-cluster distances to all cells and the average
between-cluster distances of that cell to the closest cluster divided
by their maximum.
}
\examples{
sim <- simulate_data(
  n_genes = 1000, batch_cells = c(150, 50),
  group_prob = c(0.5, 0.5), n_hvgs = 500,
  compute_pca = TRUE, output_format = "SingleCellExperiment"
)
asw <- average_silhouette_width(
  input = sim, label_true = "Group",
  reduction = "PCA"
)

}
