% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{batch_params}
\alias{batch_params}
\title{Batch params}
\usage{
batch_params(normalized, batch)
}
\arguments{
\item{normalized}{Normalized counts matrix.}

\item{batch}{A string specifying the batch variable.}
}
\value{
Batch factor location.
}
\description{
Batch params
}
