% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/batchCorrect.R
\name{batchCorrect}
\alias{batchCorrect}
\alias{batchCorrect,LimmaParams-method}
\alias{batchCorrect,CombatParams-method}
\alias{batchCorrect,SeuratV3Params-method}
\alias{batchCorrect,SeuratV5Params-method}
\alias{batchCorrect,FastMNNParams-method}
\alias{batchCorrect,HarmonyParams-method}
\alias{batchCorrect,ScanoramaParams-method}
\alias{batchCorrect,SCVIParams-method}
\alias{batchCorrect,ScMerge2Params-method}
\alias{batchCorrect,BBKNNParams-method}
\alias{batchCorrect,LigerParams-method}
\title{batchCorrect}
\usage{
batchCorrect(input, batch, params)

\S4method{batchCorrect}{LimmaParams}(input, batch, params)

\S4method{batchCorrect}{CombatParams}(input, batch, params)

\S4method{batchCorrect}{SeuratV3Params}(input, batch, params)

\S4method{batchCorrect}{SeuratV5Params}(input, batch, params)

\S4method{batchCorrect}{FastMNNParams}(input, batch, params)

\S4method{batchCorrect}{HarmonyParams}(input, batch, params)

\S4method{batchCorrect}{ScanoramaParams}(input, batch, params)

\S4method{batchCorrect}{SCVIParams}(input, batch, params)

\S4method{batchCorrect}{ScMerge2Params}(input, batch, params)

\S4method{batchCorrect}{BBKNNParams}(input, batch, params)

\S4method{batchCorrect}{LigerParams}(input, batch, params)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{batch}{A string specifying the batch variable.}

\item{params}{A \linkS4class{BatChefParams} object specifying
the batch correction method to use and the parameters for its execution.}
}
\value{
A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat}, or
`AnnData` object, where the output of the method (such as the corrected
gene expression matrix and/or the corrected dimensional reduction space) is
stored within the original input object.
}
\description{
A common interface for single-cell batch correction methods.
}
\details{
Users can pass parameters to each method via the constructors for params.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 50),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
out <- batchCorrect(input = sim, batch = "Batch",
                    params = ScanoramaParams(assay_type = "logcounts",
                                             return_dimred = TRUE))
}
