% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatChef-package.R
\docType{package}
\name{BatChef-package}
\alias{BatChef}
\alias{BatChef-package}
\title{BatChef: Single-cell RNA-seq batch effects correction methods interface}
\description{
This package implements a variety of methods for batch correction in single-cell RNA sequencing (scRNA-seq) data. It incorporates quantitative metrics (e.g. Wasserstein distance, Adjusted Rand Index) to evaluate their performance. Furthermore, the package assists users in identifying and applying the optimal method for specific datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zuinelena3/BatChef}
  \item Report bugs at \url{https://github.com/zuinelena3/BatChef/issues}
}

}
\author{
\strong{Maintainer}: Elena Zuin \email{elena.zuin.3@phd.unipd.it} (\href{https://orcid.org/0009-0006-3060-4835}{ORCID}) (Department of Biology, University of Padova)

Other contributors:
\itemize{
  \item Chiara Romualdi (Department of Biology, University of Padova) [contributor]
  \item Davide Risso (Department of Statistical Sciences, University of Padova) [contributor]
  \item Gabriele Sales (Department of Biology, University of Padova) [contributor]
}

}
\keyword{internal}
