% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatChefParams.R, R/Classes.R
\docType{class}
\name{LimmaParams}
\alias{LimmaParams}
\alias{CombatParams}
\alias{BatChefParams-class}
\alias{LimmaParams-class}
\alias{CombatParams-class}
\title{BatChefParams methods}
\usage{
LimmaParams(assay_type = "logcounts", ...)

CombatParams(assay_type = "counts", ...)
}
\arguments{
\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
Constructors and methods for the params parameter classes.
BatChefParams objects contain method specific parameters to pass to the batchCorrect generic.
}
