% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknnInput.R
\name{bbknnInput}
\alias{bbknnInput}
\alias{bbknnInput,Seurat-method}
\alias{bbknnInput,Seurat,Seurat-method}
\alias{bbknnInput,SingleCellExperiment-method}
\alias{bbknnInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{bbknnInput,AnnDataR6-method}
\alias{bbknnInput,AnnDataR6,AnnDataR6-method}
\title{Convert into a BBKNN compatible object}
\usage{
bbknnInput(input, batch, reduction)

\S4method{bbknnInput}{Seurat}(input, batch, reduction)

\S4method{bbknnInput}{SingleCellExperiment}(input, batch, reduction)

\S4method{bbknnInput}{AnnDataR6}(input, batch, reduction)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{batch}{A string specifying the batch variable.}

\item{reduction}{A string specifying the name of PCA.}
}
\value{
A BBKNN compatible object
}
\description{
Convert into a BBKNN compatible object
}
