% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknnPost.R
\name{bbknnPost}
\alias{bbknnPost}
\alias{bbknnPost,Seurat-method}
\alias{bbknnPost,Seurat,Seurat-method}
\alias{bbknnPost,SingleCellExperiment-method}
\alias{bbknnPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{bbknnPost,AnnDataR6-method}
\alias{bbknnPost,AnnDataR6,AnnDataR6-method}
\title{Convert BBKNN output}
\usage{
bbknnPost(input, output, method)

\S4method{bbknnPost}{Seurat}(input, output, method)

\S4method{bbknnPost}{SingleCellExperiment}(input, output, method)

\S4method{bbknnPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{output}{BBKNN output: a matrix of the coordinates of the points chosen
to represent the dissimilarities.}

\item{method}{A string specifying the correction method.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
\description{
Convert BBKNN output into a \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
