% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknnRun.R
\name{bbknnRun}
\alias{bbknnRun}
\title{BBKNN method}
\usage{
bbknnRun(
  input,
  batch,
  neighbors_within_batch = 3,
  n_pcs = 50,
  trim = NULL,
  computation = "annoy",
  annoy_n_trees = 10,
  pynndescent_n_neighbors = 30,
  pynndescent_random_state = 0,
  metric = "euclidean",
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  approx = NULL,
  use_annoy = NULL,
  use_faiss = NULL,
  scanpy_logging = FALSE
)
}
\arguments{
\item{input}{PCA (or other dimensionality reduction) coordinates for each cell, with cells as rows.}

\item{batch}{Vector of batch assignments for each cell.}

\item{neighbors_within_batch}{Number of many top neighbours to report for each batch.}

\item{n_pcs}{Number of dimensions.}

\item{trim}{Trim the neighbours of each cell to these many top connectivities.}

\item{computation}{A string specifying the KNN algorithm to use.}

\item{annoy_n_trees}{The number of trees to construct in the annoy forest.}

\item{pynndescent_n_neighbors}{The number of neighbours to include in the approximate neighbour graph.}

\item{pynndescent_random_state}{The RNG seed to use when creating the graph.}

\item{metric}{A string specifying the distance metric to use.}

\item{set_op_mix_ratio}{UMAP connectivity computation parameter}

\item{local_connectivity}{UMAP connectivity computation parameter,}

\item{approx}{A logical to approximate nearest neighbor method selection.}

\item{use_annoy}{Flag to explicitly enable or disable Annoy for neighbor search.}

\item{use_faiss}{Flag to explicitly enable or disable FAISS (Facebook AI Similarity Search) for neighbor search.}

\item{scanpy_logging}{If True, enables logging compatible with Scanpy’s logging system for better integration.}
}
\description{
BBKNN method
}
