% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering after integration reduction}
\usage{
clustering(input, label_true, reduction)
}
\arguments{
\item{input}{input}

\item{label_true}{label_true}

\item{reduction}{reduction}
}
\description{
Compute Leiden clustering. The resolution was chosen using the NMI metric.
}
