% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering after integration reduction}
\usage{
clustering(obj, cell_type, reduction)
}
\arguments{
\item{obj}{a SingleCellExperiment object}

\item{cell_type}{a string specifying cell-type labels}

\item{reduction}{a string specifying the dimension reduction name}
}
\description{
Compute Leiden clustering. The resolution was chosen using the NMI metric.
}
