% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatRun.R
\name{combatRun}
\alias{combatRun}
\title{ComBat method}
\usage{
combatRun(input, batch, assay_type = "counts", ...)
}
\arguments{
\item{input}{A SingleCellExperiment object.}

\item{batch}{A string specifying the batch for each cell.}

\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
ComBat method
}
