% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combatRun.R
\name{combatRun}
\alias{combatRun}
\title{ComBat method}
\usage{
combatRun(input, batch, assay_type = "counts", ...)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch for each cell.}

\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\value{
A corrected matrix
}
\description{
ComBat allows users to adjust for batch effects in datasets using an
empirical Bayes framework.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 50),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
combat <- combatRun(input = sim, batch = "Batch")

}
