% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMNNPost.R
\name{fastMNNPost}
\alias{fastMNNPost}
\alias{fastMNNPost,Seurat-method}
\alias{fastMNNPost,Seurat,Seurat-method}
\alias{fastMNNPost,SingleCellExperiment-method}
\alias{fastMNNPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{fastMNNPost,AnnDataR6-method}
\alias{fastMNNPost,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
fastMNNPost(input, output, method)

\S4method{fastMNNPost}{Seurat}(input, output, method)

\S4method{fastMNNPost}{SingleCellExperiment}(input, output, method)

\S4method{fastMNNPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}
}
\description{
Title
}
