% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMNNRun.R
\name{fastMNNRun}
\alias{fastMNNRun}
\title{fastMNN method}
\usage{
fastMNNRun(input, batch, ...)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch variable.}

\item{...}{Named arguments to pass to individual methods upon dispatch}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object
is returned where each row is a gene and each column is a cell.
This contains a corrected low-dimensional coordinates and a reconstructed
matrix in the assays slot.
}
\description{
Correct for batch effects in single-cell expression data using a fast
version of the mutual nearest neighbors (MNN) method.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
fastmnn <- fastMNNRun(input = sim, batch = "Batch")

}
