% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonyPost.R
\name{harmonyPost}
\alias{harmonyPost}
\alias{harmonyPost,Seurat-method}
\alias{harmonyPost,Seurat,Seurat-method}
\alias{harmonyPost,SingleCellExperiment-method}
\alias{harmonyPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{harmonyPost,AnnDataR6-method}
\alias{harmonyPost,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
harmonyPost(input, output, method)

\S4method{harmonyPost}{Seurat}(input, output, method)

\S4method{harmonyPost}{SingleCellExperiment}(input, output, method)

\S4method{harmonyPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}
}
\description{
Title
}
