% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonyPost.R
\name{harmonyPost}
\alias{harmonyPost}
\alias{harmonyPost,Seurat-method}
\alias{harmonyPost,Seurat,Seurat-method}
\alias{harmonyPost,SingleCellExperiment-method}
\alias{harmonyPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{harmonyPost,AnnDataR6-method}
\alias{harmonyPost,AnnDataR6,AnnDataR6-method}
\title{Convert Harmony output}
\usage{
harmonyPost(input, output, method)

\S4method{harmonyPost}{Seurat}(input, output, method)

\S4method{harmonyPost}{SingleCellExperiment}(input, output, method)

\S4method{harmonyPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{output}{Harmony output: a \link[SingleCellExperiment]{SingleCellExperiment}
containing the corrected dimensionality reduction space.}

\item{method}{A string specifying the correction method.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
\description{
Convert Harmony output into a \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
