% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonyRun.R
\name{harmonyRun}
\alias{harmonyRun}
\title{Harmony method}
\usage{
harmonyRun(input, batch, ...)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch variable.}

\item{...}{Named arguments to pass to individual methods upon dispatch}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} object which
contains a corrected low-dimensional space.
}
\description{
Harmony is a mixture-model based method.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
harmony <- harmonyRun(input = sim, batch = "Batch")

}
