% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerInput.R
\name{ligerInput}
\alias{ligerInput}
\alias{ligerInput,Seurat-method}
\alias{ligerInput,Seurat,Seurat-method}
\alias{ligerInput,SingleCellExperiment-method}
\alias{ligerInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{ligerInput,AnnDataR6-method}
\alias{ligerInput,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
ligerInput(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)

\S4method{ligerInput}{Seurat}(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)

\S4method{ligerInput}{SingleCellExperiment}(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)

\S4method{ligerInput}{AnnDataR6}(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{features}{features}

\item{useDatasets}{useDatasets}

\item{verbose}{verbose}

\item{format.type}{format.type}

\item{remove.missing}{remove.missing}

\item{...}{...}
}
\description{
Title
}
