% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerInput.R
\name{ligerInput}
\alias{ligerInput}
\alias{ligerInput,Seurat-method}
\alias{ligerInput,Seurat,Seurat-method}
\alias{ligerInput,SingleCellExperiment-method}
\alias{ligerInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{ligerInput,AnnDataR6-method}
\alias{ligerInput,AnnDataR6,AnnDataR6-method}
\title{Convert into a LIGER object}
\usage{
ligerInput(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)

\S4method{ligerInput}{Seurat}(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)

\S4method{ligerInput}{SingleCellExperiment}(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)

\S4method{ligerInput}{AnnDataR6}(
  input,
  batch,
  features,
  useDatasets = NULL,
  verbose = TRUE,
  format.type = NULL,
  remove.missing = NULL,
  ...
)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{batch}{A string specifying the batch variable.}

\item{features}{Vector of features to use.}

\item{useDatasets}{A character vector of the names, a numeric or logical vector of the index of the datasets to be normalized.}

\item{verbose}{Print progress bar/messages}

\item{format.type}{Deprecated. The functionality of these is covered through other parts of the whole workflow and is no long needed}

\item{remove.missing}{Deprecated. The functionality of these is covered through other parts of the whole workflow and is no long needed}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
Convert into a LIGER object
}
