% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerRun.R
\name{ligerRun}
\alias{ligerRun}
\title{LIGER method}
\usage{
ligerRun(
  input,
  method = "iNMF",
  quantiles = 50,
  reference = NULL,
  minCells = 20,
  nNeighbors = 20,
  useDims = NULL,
  center = FALSE,
  maxSample = 1000,
  eps = 0.9,
  refineKNN = TRUE,
  clusterName = "quantileNorm_cluster",
  seed = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{input}{A `liger` object.}

\item{method}{iNMF variant algorithm to use for integration.}

\item{quantiles}{Number of quantiles to use for quantile normalization.}

\item{reference}{Character, numeric or logical selection of one dataset, out of all available datasets in object, to use as a "reference" for quantile normalization.}

\item{minCells}{Minimum number of cells to consider a cluster shared across datasets}

\item{nNeighbors}{Number of nearest neighbors for within-dataset knn graph.}

\item{useDims}{Indices of factors to use for shared nearest factor determination.}

\item{center}{A logical to center data.}

\item{maxSample}{Maximum number of cells used for quantile normalization of each cluster and factor}

\item{eps}{The error bound of the nearest neighbor search.}

\item{refineKNN}{A logical to increase robustness of cluster assignments using KNN graph.}

\item{clusterName}{Variable name that will store the clustering result in metadata of a liger object or a Seurat object.}

\item{seed}{Random seed}

\item{verbose}{Print progress bar/messages}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
LIGER method
}
