% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerRun.R
\name{ligerRun}
\alias{ligerRun}
\title{LIGER method}
\usage{
ligerRun(
  input,
  method = "iNMF",
  quantiles = 50,
  reference = NULL,
  minCells = 20,
  nNeighbors = 20,
  useDims = NULL,
  center = FALSE,
  maxSample = 1000,
  eps = 0.9,
  refineKNN = TRUE,
  clusterName = "quantileNorm_cluster",
  seed = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{input}{input}

\item{method}{method}

\item{quantiles}{quantiles}

\item{reference}{reference}

\item{minCells}{minCells}

\item{nNeighbors}{nNeighbors}

\item{useDims}{useDims}

\item{center}{center}

\item{maxSample}{maxSample}

\item{eps}{eps}

\item{refineKNN}{refineKNN}

\item{clusterName}{clusterName}

\item{seed}{seed}

\item{verbose}{verbose}

\item{...}{...}
}
\description{
LIGER method
}
