% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerRun.R
\name{ligerRun}
\alias{ligerRun}
\title{LIGER method}
\usage{
ligerRun(
  input,
  method = "iNMF",
  quantiles = 50,
  reference = NULL,
  min_cells = 20,
  n_neighbors = 20,
  use_dims = NULL,
  center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine_knn = TRUE,
  cluster_name = "quantileNorm_cluster",
  seed = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{input}{A \linkS4class{liger} object.}

\item{method}{A string specifyng the batch correction method. Choose from
\code{"iNMF"}, \code{"onlineINMF"}, \code{"UINMF"}. Default \code{"iNMF"}.}

\item{quantiles}{Number of quantiles to use for quantile normalization.}

\item{reference}{Character, numeric or logical selection of one dataset,
out of all available datasets in object, to use as a "reference" for
quantile normalization.}

\item{min_cells}{Minimum number of cells to consider a cluster shared across
datasets}

\item{n_neighbors}{Number of nearest neighbors for within-dataset knn graph.}

\item{use_dims}{Indices of factors to use for shared nearest factor
determination.}

\item{center}{A logical to center data.}

\item{max_sample}{Maximum number of cells used for quantile normalization of
each cluster and factor}

\item{eps}{The error bound of the nearest neighbor search.}

\item{refine_knn}{A logical to increase robustness of cluster assignments
using KNN graph.}

\item{cluster_name}{Variable name that will store the clustering result
in metadata of a liger object or a Seurat object.}

\item{seed}{Random seed}

\item{verbose}{Print progress bar/messages}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\value{
A \linkS4class{liger} object, which contains the quantile align
factor loadings.
}
\description{
LIGER is a integrative non-negative matrix factorization method
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                    compute_pca = TRUE, output_format = "SingleCellExperiment")

ll <- lapply(unique(SingleCellExperiment::colData(sim)[, "Batch"]),
             function(i) sim[, SingleCellExperiment::colData(sim)[, "Batch"] == i])
names(ll) <- unique(SingleCellExperiment::colData(sim)[, "Batch"])
# Create a liger object
lo <- rliger::createLiger(rawData = ll)
lo <- rliger::normalize(object = lo)
lo <- rliger::scaleNotCenter(object = lo, features = rownames(sim))
liger <- ligerRun(input = lo)

}
