% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaRun.R
\name{limmaRun}
\alias{limmaRun}
\title{limma method}
\usage{
limmaRun(input, batch, assay_type = "logcounts", ...)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch variable.}

\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\value{
A corrected matrix
}
\description{
limma method
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 200),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
limma <- limmaRun(input = sim, batch = "Batch")

}
