% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaRun.R
\name{limmaRun}
\alias{limmaRun}
\title{limma method}
\usage{
limmaRun(input, batch, assay_type = "logcounts", ...)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{batch}{A string specifying the batch variable.}

\item{assay_type}{A string specifying the assay.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\value{
A corrected matrix
}
\description{
limma method
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = FALSE,
                     output_format = "SingleCellExperiment")
limma <- limmaRun(input = sim, batch = "Batch")

}
