% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligerPost.R, R/linearPost.R
\name{linearPost}
\alias{linearPost}
\alias{linearPost,Seurat-method}
\alias{linearPost,Seurat,Seurat-method}
\alias{linearPost,SingleCellExperiment-method}
\alias{linearPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{linearPost,AnnDataR6-method}
\alias{linearPost,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
linearPost(input, output, method)

\S4method{linearPost}{Seurat}(input, output, method)

\S4method{linearPost}{SingleCellExperiment}(input, output, method)

\S4method{linearPost}{AnnDataR6}(input, output, method)

linearPost(input, output, method)

\S4method{linearPost}{Seurat}(input, output, method)

\S4method{linearPost}{SingleCellExperiment}(input, output, method)

\S4method{linearPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}
}
\description{
Title

Title
}
