% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_inverse_simpson_index.R
\name{local_inverse_simpson_index}
\alias{local_inverse_simpson_index}
\title{Local Inverse Simposon Index}
\usage{
local_inverse_simpson_index(
  input,
  label_true,
  reduction,
  meta_data = colData(input),
  perplexity = 30,
  nn_eps = 0
)
}
\arguments{
\item{input}{A `SingleCellExperiment` object.}

\item{label_true}{A string specifying cell types.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{meta_data}{A data frame with one row per cell.}

\item{perplexity}{The effective number of each cell's neighbors.}

\item{nn_eps}{Error bound for nearest neighbor search with `RANN:nn2()`.}
}
\description{
Local Inverse Simposon Index
}
