% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_inverse_simpson_index.R
\name{local_inverse_simpson_index}
\alias{local_inverse_simpson_index}
\title{Local Inverse Simposon Index (LISI)}
\usage{
local_inverse_simpson_index(
  input,
  label_true,
  reduction,
  meta_data = colData(input),
  perplexity = 30,
  nn_eps = 0
)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment} object.}

\item{label_true}{A string specifying cell types.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{meta_data}{A data frame with one row per cell.}

\item{perplexity}{The effective number of each cell's neighbors.}

\item{nn_eps}{Error bound for nearest neighbor search with `RANN:nn2()`.}
}
\description{
Local Inverse Simpson’s Index (LISI) is a local level metric based on
the kNN algorithm. This metric defines the effective number of datasets in
a neighborhood. In other words, LISI determines the number of neighbor cells
necessary before one batch is observed twice.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 200),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
lisi <- local_inverse_simpson_index(input = sim, label_true = "Group",
                                    reduction = "PCA")

}
