% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Common quantitative metrics}
\usage{
metrics(anndata, cell_type, batch, reduction, only_ARI = FALSE)
}
\arguments{
\item{anndata}{anndata object, computed by clustering function}

\item{cell_type}{string specifying cell-type labels}

\item{batch}{string specifying batch}

\item{reduction}{string specifying the integration reduction}

\item{only_ARI}{compute only ARI metrics}
}
\description{
Common quantitative metrics
}
