% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Evaluation metrics function}
\usage{
metrics(
  input,
  batch,
  group,
  reduction,
  rep = 10,
  mc.cores = 2,
  average_method = "arithmetic",
  metric = "euclidean",
  sample_size = NULL,
  random_state = NULL,
  meta_data = colData(input),
  perplexity = 30,
  nn_eps = 0
)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{group}{group}

\item{reduction}{reduction}

\item{rep}{rep}

\item{mc.cores}{mc.cores}

\item{average_method}{average_method}

\item{metric}{metric}

\item{sample_size}{sample_size}

\item{random_state}{random_state}

\item{meta_data}{meta_data}

\item{perplexity}{perplexity}

\item{nn_eps}{nn_eps}
}
\description{
Evaluation metrics function
}
