% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_mutual_info.R
\name{normalized_mutual_info}
\alias{normalized_mutual_info}
\title{Normalized Mutual Information (NMI)}
\usage{
normalized_mutual_info(
  input,
  label_true,
  average_method = "arithmetic",
  reduction
)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{label_true}{A string specifying cell types.}

\item{average_method}{How to compute the normalizer in the denominator.}

\item{reduction}{A string specifying the dimensional reduction.}
}
\description{
Before NMI, the Leiden clustering is computed.
}
