% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_mutual_info.R
\name{normalized_mutual_info}
\alias{normalized_mutual_info}
\title{Normalized Mutual Information (NMI)}
\usage{
normalized_mutual_info(
  input,
  label_true,
  reduction,
  nmi_compute = FALSE,
  resolution,
  average_method = "arithmetic"
)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{label_true}{A string specifying the ground truth labels.}

\item{reduction}{A string specifying the dimensional reduction
on which the clustering analysis will be performed.}

\item{nmi_compute}{A Boolean value indicating NMI metric calculation to
identify the optimal clustering is to be performed (Default: FALSE).}

\item{resolution}{A numeric value specifying the resolution parameter.}

\item{average_method}{How to compute the normalizer in the denominator.}
}
\value{
A numeric value.
}
\description{
The Normalized Mutual Information (NMI) metric is used to compare
the overlap between the true cell-type and clustering labels computed after
batch correction.
}
\details{
After computing clustering using Leiden algorithm, the NMI metric is computed
using the `normalized_mutual_info_score` function from the Python sklearn package.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
nmi <- normalized_mutual_info(input = sim, label_true = "Group", reduction = "PCA",
                           nmi_compute = FALSE, resolution = 0.5)

}
