% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_mutual_info.R
\name{normalized_mutual_info}
\alias{normalized_mutual_info}
\title{Normalized Mutual Information (NMI)}
\usage{
normalized_mutual_info(
  input,
  label_true,
  average_method = "arithmetic",
  reduction
)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{label_true}{A string specifying cell types.}

\item{average_method}{How to compute the normalizer in the denominator.}

\item{reduction}{A string specifying the dimensional reduction.}
}
\value{
A numeric value.
}
\description{
The Normalized Mutual Information (NMI) metric is used to compare
the overlap between the true cell-type and clustering labels computed after
batch correction.
}
\details{
After computing clustering using Leiden algorithm, the ARI metric is computed
using the `normalized_mutual_info_score` function from the Python sklearn package.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(150, 50),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = TRUE, output_format = "SingleCellExperiment")
nmi <- normalized_mutual_info(input = sim, label_true = "Group",
                              reduction = "PCA")

}
