% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{outlier_params}
\alias{outlier_params}
\title{Expression outlier parameters}
\usage{
outlier_params(normalized)
}
\arguments{
\item{counts}{Normalized counts matrix.}
}
\value{
Probability that genes will be selected to be expression outliers.
}
\description{
Expression outlier parameters
}
