% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaInput.R
\name{scanoramaInput}
\alias{scanoramaInput}
\alias{scanoramaInput,Seurat-method}
\alias{anndatasInput,Seurat,Seurat-method}
\alias{scanoramaInput,SingleCellExperiment-method}
\alias{anndatasInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{scanoramaInput,AnnDataR6-method}
\alias{anndatasInput,AnnDataR6,AnnDataR6-method}
\title{Convert to a Scanorama compatible object}
\usage{
scanoramaInput(input, batch, assay.type = NULL)

\S4method{scanoramaInput}{Seurat}(input, batch, assay.type = NULL)

\S4method{scanoramaInput}{SingleCellExperiment}(input, batch, assay.type = NULL)

\S4method{scanoramaInput}{AnnDataR6}(input, batch, assay.type = NULL)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{batch}{A string specifying the batch for each cell.}

\item{assay.type}{layers name}
}
\value{
List that contains the expression matrices for each batches and
 the genes.
}
\description{
Convert to a Scanorama compatible object
}
