% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaPost.R
\name{scanoramaPost}
\alias{scanoramaPost}
\alias{scanoramaPost,Seurat-method}
\alias{scanoramaPost,Seurat,Seurat-method}
\alias{scanoramaPost,SingleCellExperiment-method}
\alias{scanoramaPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{scanoramaPost,AnnDataR6-method}
\alias{scanoramaPost,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
scanoramaPost(input, output, batch, return_dimred, method)

\S4method{scanoramaPost}{Seurat}(input, output, batch, return_dimred, method)

\S4method{scanoramaPost}{SingleCellExperiment}(input, output, batch, return_dimred, method)

\S4method{scanoramaPost}{AnnDataR6}(input, output, batch, return_dimred, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{batch}{batch}

\item{return_dimred}{return_dimred}

\item{method}{method}
}
\description{
Title
}
