% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaPost.R
\name{scanoramaPost}
\alias{scanoramaPost}
\alias{scanoramaPost,Seurat-method}
\alias{scanoramaPost,Seurat,Seurat-method}
\alias{scanoramaPost,SingleCellExperiment-method}
\alias{scanoramaPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{scanoramaPost,AnnDataR6-method}
\alias{scanoramaPost,AnnDataR6,AnnDataR6-method}
\title{Convert the Scanorama output}
\usage{
scanoramaPost(input, list, output, return_dimred, method)

\S4method{scanoramaPost}{Seurat}(input, list, output, return_dimred, method)

\S4method{scanoramaPost}{SingleCellExperiment}(input, list, output, return_dimred, method)

\S4method{scanoramaPost}{AnnDataR6}(input, list, output, return_dimred, method)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{list}{List version of input.}

\item{output}{Scanorama output: a list which contains the transposed corrected
gene expression matrix and the corrected low-dimensional space.}

\item{return_dimred}{A logical to returning integrated low-dimesional embeddings.}

\item{method}{A string specifying the correction method}
}
\description{
Convert the Scanorama output into a \linkS4class{SingleCellExperiment},
\linkS4class{Seurat} or `AnnData` object.
}
