% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaRun.R
\name{scanoramaRun}
\alias{scanoramaRun}
\title{scanorama method}
\usage{
scanoramaRun(
  input,
  return_dimred = FALSE,
  batch_size = 5000,
  verbose = TRUE,
  ds_names = NULL,
  dimred = 100,
  approx = TRUE,
  sigma = 15,
  alpha = 0.1,
  knn = 20,
  return_dense = FALSE,
  hvg = NULL,
  union = FALSE,
  seed = 0
)
}
\arguments{
\item{input}{input}

\item{return_dimred}{return_dimred}

\item{batch_size}{batch_size}

\item{verbose}{verbose}

\item{ds_names}{ds_names}

\item{dimred}{dimred}

\item{approx}{approx}

\item{sigma}{sigma}

\item{alpha}{alpha}

\item{knn}{knn}

\item{return_dense}{return_dense}

\item{hvg}{hvg}

\item{union}{union}

\item{seed}{seed}
}
\description{
scanorama method
}
