% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanoramaRun.R
\name{scanoramaRun}
\alias{scanoramaRun}
\title{Scanorama method}
\usage{
scanoramaRun(
  input,
  return_dimred = FALSE,
  batch_size = 5000,
  verbose = TRUE,
  ds_names = NULL,
  dimred = 100,
  approx = TRUE,
  sigma = 15,
  alpha = 0.1,
  knn = 20,
  return_dense = FALSE,
  hvg = NULL,
  union = FALSE,
  seed = 0
)
}
\arguments{
\item{input}{List that contains the expression matrices for each batches and
the genes.}

\item{return_dimred}{A logical to returning integrated low-dimesional
embeddings.}

\item{batch_size}{The batch size used in the alignment vector computation.}

\item{verbose}{Print progress bars and output.}

\item{ds_names}{Reports data set names in logging output}

\item{dimred}{Number of dimensions of corrected embedding.}

\item{approx}{A logical to use approximate nearest neighbors}

\item{sigma}{Correction smoothing parameter on Gaussian kernel.}

\item{alpha}{Alignment score minimum cutoff.}

\item{knn}{Number of nearest neighbors to use for matching.}

\item{return_dense}{A logical to return dense matrices.}

\item{hvg}{Use this number of top highly variable genes based on dispersion.}

\item{union}{A logical to consider the union of the genes.
In alternative, the intersection is considered.}

\item{seed}{Random seed to use.}
}
\value{
A list that contains the corrected gene expression matrix and
the corrected low-dimensional reduction
}
\description{
Scanorama is a Mutual Nearest Neighbors method. This an algorithm that
identifies and merges the shared cell types among all pairs of datasets and
accurately integrates heterogeneous collections of scRNA-seq data.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 200),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
sim <- lapply(unique(colData(sim)[, "Batch"]), function(x) sim[, colData(sim)[, "Batch"] == x])
ll <- c(lapply(sim, function(x) t(as.matrix(logcounts(x)))),
        lapply(sim, function(x) rownames(x)))
scanorama <- scanoramaRun(input = ll, return_dimred = TRUE)

}
