% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceInput.R
\name{sceInput}
\alias{sceInput}
\alias{sceInput,Seurat-method}
\alias{sceInput,Seurat,Seurat-method}
\alias{sceInput,SingleCellExperiment-method}
\alias{sceInput,SingleCellExperiment,SingleCellExperiment-method}
\alias{sceInput,AnnDataR6-method}
\alias{sceInput,AnnDataR6,AnnDataR6-method}
\title{Convert into a SingleCellExperiment object}
\usage{
sceInput(input, batch)

\S4method{sceInput}{Seurat}(input, batch)

\S4method{sceInput}{SingleCellExperiment}(input, batch)

\S4method{sceInput}{AnnDataR6}(input, batch)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{batch}{A string specifying the batch for each cell.}
}
\description{
Convert into a SingleCellExperiment object
}
