% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMerge2Post.R
\name{scMerge2Post}
\alias{scMerge2Post}
\alias{scMerge2Post,Seurat-method}
\alias{scMerge2Post,Seurat,Seurat-method}
\alias{scMerge2Post,SingleCellExperiment-method}
\alias{scMerge2Post,SingleCellExperiment,SingleCellExperiment-method}
\alias{scMerge2Post,AnnDataR6-method}
\alias{scMerge2Post,AnnDataR6,AnnDataR6-method}
\title{Convert the scMerge2 output}
\usage{
scMerge2Post(input, output, method)

\S4method{scMerge2Post}{Seurat}(input, output, method)

\S4method{scMerge2Post}{SingleCellExperiment}(input, output, method)

\S4method{scMerge2Post}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{output}{scMerge2 output: a list that contains the corrected
gene expression matrix}

\item{method}{A string specifying the correction method}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
\description{
Convert the scMerge2 output into a \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
