% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMerge2Run.R
\name{scMerge2Run}
\alias{scMerge2Run}
\title{scMerge2 method}
\usage{
scMerge2Run(input, batch, assay_type = "logcounts", ...)
}
\arguments{
\item{input}{A `SingleCellExperiment` object can be supplied.}

\item{batch}{A string specifying the batch variable.}

\item{assay_type}{A string specifying the assay to use for correction.}

\item{...}{Named arguments to pass to individual methods upon dispatch.}
}
\description{
scMerge2 method
}
