% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVIPost.R
\name{scVIPost}
\alias{scVIPost}
\alias{scVIPost,Seurat-method}
\alias{scVIPost,Seurat,Seurat-method}
\alias{scVIPost,SingleCellExperiment-method}
\alias{scVIPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{scVIPost,AnnDataR6-method}
\alias{scVIPost,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
scVIPost(input, output, method)

\S4method{scVIPost}{Seurat}(input, output, method)

\S4method{scVIPost}{SingleCellExperiment}(input, output, method)

\S4method{scVIPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{method}{method}
}
\description{
Title
}
