% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVIPost.R
\name{scVIPost}
\alias{scVIPost}
\alias{scVIPost,Seurat-method}
\alias{scVIPost,Seurat,Seurat-method}
\alias{scVIPost,SingleCellExperiment-method}
\alias{scVIPost,SingleCellExperiment,SingleCellExperiment-method}
\alias{scVIPost,AnnDataR6-method}
\alias{scVIPost,AnnDataR6,AnnDataR6-method}
\title{Convert the scVI output}
\usage{
scVIPost(input, output, method)

\S4method{scVIPost}{Seurat}(input, output, method)

\S4method{scVIPost}{SingleCellExperiment}(input, output, method)

\S4method{scVIPost}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object can be supplied.}

\item{output}{scVI output: a list that contains the corrected
gene expression matrix and the corrected low-dimensional space.}

\item{method}{A string specifying the correction method}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
\description{
Convert the scVI output into a \link[SingleCellExperiment]{SingleCellExperiment}
\link[Seurat]{Seurat} or `AnnData` object.
}
