% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratv3Post.R
\name{seuratv3Post}
\alias{seuratv3Post}
\alias{seuratv3Post,Seurat-method}
\alias{seuratv3Post,Seurat,Seurat-method}
\alias{seuratv3Post,SingleCellExperiment-method}
\alias{seuratv3Post,SingleCellExperiment,SingleCellExperiment-method}
\alias{seuratv3Post,list-method}
\alias{seuratv3Post,AnnDataR6,AnnDataR6-method}
\title{Convert the Seurat V3 output}
\usage{
seuratv3Post(input, output, method)

\S4method{seuratv3Post}{Seurat}(input, output, method)

\S4method{seuratv3Post}{SingleCellExperiment}(input, output, method)

\S4method{seuratv3Post}{list}(input, output, method)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{output}{Seurat V3 output}

\item{method}{A string specifying the correction method}
}
\description{
Convert the Seurat V3 output into a SingleCellExperiment, Seurat or Anndata objects
}
