% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratv3Post.R
\name{seuratv3Post}
\alias{seuratv3Post}
\alias{seuratv3Post,Seurat-method}
\alias{seuratv3Post,Seurat,Seurat-method}
\alias{seuratv3Post,SingleCellExperiment-method}
\alias{seuratv3Post,SingleCellExperiment,SingleCellExperiment-method}
\alias{seuratv3Post,AnnDataR6-method}
\alias{seuratv3Post,AnnDataR6,AnnDataR6-method}
\title{Convert the SeuratV3 output}
\usage{
seuratv3Post(input, output, method)

\S4method{seuratv3Post}{Seurat}(input, output, method)

\S4method{seuratv3Post}{SingleCellExperiment}(input, output, method)

\S4method{seuratv3Post}{AnnDataR6}(input, output, method)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment}, \linkS4class{Seurat} or
`AnnData` object can be supplied.}

\item{output}{Seurat V3 output: a Seurat object}

\item{method}{A string specifying the correction method}
}
\value{
A \linkS4class{SingleCellExperiment},
\linkS4class{Seurat} or `AnnData` object
}
\description{
Convert the SeuratV3 output into a \linkS4class{SingleCellExperiment},
\linkS4class{Seurat} or `AnnData` object
}
