% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratV3Run.R
\name{seuratV3Run}
\alias{seuratV3Run}
\title{Seurat V3 method}
\usage{
seuratV3Run(
  input,
  assay = NULL,
  reference = NULL,
  anchor_features = 2000,
  scale = TRUE,
  normalization_method = "LogNormalize",
  sct_clip_range = NULL,
  reduction = "cca",
  l2_norm = TRUE,
  dims = 1:30,
  k_anchor = 5,
  k_filter = 200,
  k_score = 30,
  max_features = 200,
  nn_method = "annoy",
  n_trees = 50,
  eps = 0,
  verbose = FALSE,
  new_assay_name = "integrated",
  features = NULL,
  features_to_integrate = NULL,
  k_weight = 100,
  weight_reduction = NULL,
  sd_weight = 1,
  sample_tree = NULL,
  preserve_order = FALSE
)
}
\arguments{
\item{input}{A list of \link[Seurat]{Seurat} objects.}

\item{assay}{A vector of assay names specifying which assay to use when
constructing anchors.}

\item{reference}{A vector specifying the object/s to be used as a reference
during integration.}

\item{anchor_features}{Number of features to be used in anchor finding.}

\item{scale}{A logical to scale the features provided.}

\item{normalization_method}{Name of normalization method used:
LogNormalize (default) or SCT.}

\item{sct_clip_range}{Numeric of length two specifying the min and max values
the Pearson residual will be clipped to.}

\item{reduction}{Dimensional reduction to perform when finding anchors.}

\item{l2_norm}{Perform L2 normalization on the CCA cell embeddings
after dimensional reduction.}

\item{dims}{Number of dimensions.}

\item{k_anchor}{Number of neighbors (k) to use when picking anchors.}

\item{k_filter}{Number of neighbors (k) to use when filtering anchors.}

\item{k_score}{Number of neighbors (k) to use when scoring anchors.}

\item{max_features}{The maximum number of features to use when specifying
the neighborhood search space in the anchor filtering.}

\item{nn_method}{Method for nearest neighbor finding.}

\item{n_trees}{More trees gives higher precision when using annoy
approximate nearest neighbor search.}

\item{eps}{Error bound on the neighbor finding algorithm.}

\item{verbose}{Print progress bars and output.}

\item{new_assay_name}{Name for the new assay containing
the integrated data.}

\item{features}{Vector of features to use.}

\item{features_to_integrate}{Vector of features to integrate.}

\item{k_weight}{Number of neighbors to consider when weighting anchors.}

\item{weight_reduction}{Dimension reduction to use when calculating
anchor weights.}

\item{sd_weight}{Controls the bandwidth of the Gaussian kernel for weighting.}

\item{sample_tree}{Specify the order of integration.}

\item{preserve_order}{Do not reorder objects based on size for each
pairwise integration.}
}
\value{
A \link[Seurat]{Seurat} object that contains the corrected matrix.
}
\description{
SeuratV3 is an anchor-based method.
}
\examples{
sim <- simulate_data(n_genes = 1000, batch_cells = c(200, 200),
                     group_prob = c(0.5, 0.5), n_hvgs = 500,
                     compute_pca = FALSE,
                     output_format = "Seurat")
feat <- Seurat::VariableFeatures(sim)
sim <- Seurat::SplitObject(sim, split.by = "Batch")
seuv3 <- seuratV3Run(input = sim, reduction = "cca",
                     features = feat)

}
