% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratV5Input.R
\name{seuratv5Input}
\alias{seuratv5Input}
\alias{seuratv5Input,Seurat-method}
\alias{seuratv5Input,Seurat,Seurat-method}
\alias{seuratv5Input,SingleCellExperiment-method}
\alias{seuratv5Input,SingleCellExperiment,SingleCellExperiment-method}
\alias{seuratv5Input,AnnDataR6-method}
\alias{seuratv5Input,AnnDataR6,AnnDataR6-method}
\title{Title}
\usage{
seuratv5Input(input, batch, pca_name = NULL)

\S4method{seuratv5Input}{Seurat}(input, batch, pca_name = NULL)

\S4method{seuratv5Input}{SingleCellExperiment}(input, batch, pca_name = NULL)

\S4method{seuratv5Input}{AnnDataR6}(input, batch, pca_name = NULL)
}
\arguments{
\item{input}{input}

\item{batch}{batch}

\item{pca_name}{pca_name}
}
\description{
Title
}
