% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratV5Post.R
\name{seuratv5Post}
\alias{seuratv5Post}
\alias{seuratv5Post,Seurat-method}
\alias{seuratv5Post,Seurat,Seurat-method}
\alias{seuratv5Post,SingleCellExperiment-method}
\alias{seuratv5Post,SingleCellExperiment,SingleCellExperiment-method}
\alias{seuratv5Post,AnnDataR6-method}
\alias{seuratv5Post,AnnDataR6,AnnDataR6-method}
\title{Convert the Seurat V5 output}
\usage{
seuratv5Post(input, output, method, name)

\S4method{seuratv5Post}{Seurat}(input, output, method, name)

\S4method{seuratv5Post}{SingleCellExperiment}(input, output, method, name)

\S4method{seuratv5Post}{AnnDataR6}(input, output, method, name)
}
\arguments{
\item{input}{A `SingleCellExperiment`, `Seurat` or `AnnData` objects can be supplied.}

\item{output}{Seurat V5 output.}

\item{method}{A string specifying the correction method.}

\item{name}{A string specifying the corrected reduce space name.}
}
\description{
Convert the Seurat V5 output into a SingleCellExperiment, Seurat or Anndata objects
}
