% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratV5Run.R
\name{seuratV5Run}
\alias{seuratV5Run}
\title{Seurat V5 method}
\usage{
seuratV5Run(
  input,
  method = "CCAIntegration",
  orig.reduction = "pca",
  assay = NULL,
  features = NULL,
  layers = NULL,
  scale.layer = "scale.data",
  new.reduction = "integrated.dr",
  reference = NULL,
  normalization.method = "LogNormalize",
  dims = 1:30,
  k.filter = NA,
  dims.to.integrate = NULL,
  k.weight = 100,
  weight.reduction = NULL,
  sd.weight = 1,
  sample.tree = NULL,
  preserve.order = FALSE,
  verbose = TRUE,
  l2.norm = TRUE,
  k.anchor = 5,
  k.score = 30,
  max.features = 200,
  nn.method = "annoy",
  n.trees = 50,
  eps = 0
)
}
\arguments{
\item{input}{input}

\item{method}{method}

\item{orig.reduction}{orig.reduction}

\item{assay}{assay}

\item{features}{features}

\item{layers}{layers}

\item{scale.layer}{scale.layer}

\item{new.reduction}{new.reduction}

\item{reference}{reference}

\item{normalization.method}{normalization.method}

\item{dims}{dims}

\item{k.filter}{k.filter}

\item{dims.to.integrate}{dims.to.integrate}

\item{k.weight}{k.weight}

\item{weight.reduction}{weight.reduction}

\item{sd.weight}{sd.weight}

\item{sample.tree}{sample.tree}

\item{preserve.order}{preserve.order}

\item{verbose}{verbose}

\item{l2.norm}{l2.norm}

\item{k.anchor}{k.anchor}

\item{k.score}{k.score}

\item{max.features}{max.features}

\item{nn.method}{nn.method}

\item{n.trees}{n.trees}

\item{eps}{eps}
}
\description{
Seurat V5 method
}
