% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratV5Run.R
\name{seuratV5Run}
\alias{seuratV5Run}
\title{Seurat V5 method}
\usage{
seuratV5Run(
  input,
  method = "CCAIntegration",
  orig_reduction = "pca",
  assay = NULL,
  features = NULL,
  layers = NULL,
  scale_layer = "scale.data",
  new_reduction = "integrated.dr",
  reference = NULL,
  normalization_method = "LogNormalize",
  dims = 1:30,
  k_filter = NA,
  dims_to_integrate = NULL,
  k_weight = 100,
  weight_reduction = NULL,
  sd_weight = 1,
  sample_tree = NULL,
  preserve_order = FALSE,
  verbose = FALSE,
  l2_norm = TRUE,
  k_anchor = 5,
  k_score = 30,
  max_features = 200,
  nn_method = "annoy",
  n_trees = 50,
  eps = 0
)
}
\arguments{
\item{input}{A \link[Seurat]{Seurat} object.}

\item{method}{Integration method function.}

\item{orig_reduction}{Name of dimensional reduction for correction.}

\item{assay}{Name of assay for integration.}

\item{features}{A vector of features to use for integration.}

\item{layers}{Names of normalized layers in assay.}

\item{scale_layer}{Name(s) of scaled layer(s) in assay.}

\item{new_reduction}{Name of new integrated dimensional reduction.}

\item{reference}{A reference Seurat object.}

\item{normalization_method}{Name of normalization method used:
LogNormalize or SCT.}

\item{dims}{Number of dimensions of dimensional reduction.}

\item{k_filter}{Number of anchors to filter.}

\item{dims_to_integrate}{Number of dimensions to return integrated values for.}

\item{k_weight}{Number of neighbors to consider when weighting anchors.}

\item{weight_reduction}{Dimension reduction to use when calculating
anchor weights.}

\item{sd_weight}{Controls the bandwidth of the Gaussian kernel for weighting.}

\item{sample_tree}{Specify the order of integration.}

\item{preserve_order}{Do not reorder objects based on size for each
pairwise integration.}

\item{verbose}{Print progress bars and output.}

\item{l2_norm}{Perform L2 normalization on the CCA cell embeddings after
dimensional reduction.}

\item{k_anchor}{Number of neighbors (k) to use when picking anchors.}

\item{k_score}{Number of neighbors (k) to use when scoring anchors.}

\item{max_features}{The maximum number of features to use when specifying
the neighborhood search space in the anchor filtering.}

\item{nn_method}{Method for nearest neighbor finding.}

\item{n_trees}{More trees gives higher precision when using annoy approximate
nearest neighbor search.}

\item{eps}{Error bound on the neighbor finding algorithm.}
}
\value{
A \link[Seurat]{Seurat} object.
}
\description{
SeuratV5 is an anchor-based method.
}
\examples{
sim <- simulate_data(
  n_genes = 1000, batch_cells = c(250, 200),
  group_prob = c(0.5, 0.5), n_hvgs = 500,
  compute_pca = TRUE,
  output_format = "Seurat"
)
sim[[SeuratObject::DefaultAssay(sim)]] <- split(
  x = sim[[SeuratObject::DefaultAssay(sim)]],
  f = sim[["Batch"]][, 1]
)
sim <- Seurat::ScaleData(sim, verbose = FALSE)
seuv5 <- seuratV5Run(
  input = sim, method = "CCAIntegration",
  features = rownames(sim)
)

}
