% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{simulated_data}
\alias{simulated_data}
\title{Simulated data (with splatter package)}
\usage{
simulated_data(nGenes, batchCells, group.prob, ncomp)
}
\arguments{
\item{nGenes}{number of genes}

\item{batchCells}{number of cells for each batch}

\item{group.prob}{percent of cell types}

\item{ncomp}{number of principal components}
}
\description{
Simulated data (with splatter package)
}
