% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{simulated_data}
\alias{simulated_data}
\title{Function to simulate data}
\usage{
simulated_data(
  nGenes = 10000,
  batchCells = 100,
  batch.facLoc = 0.1,
  batch.facScale = 0.1,
  batch.rmEffect = FALSE,
  mean.rate = 0.3,
  mean.shape = 0.6,
  lib.loc = 11,
  lib.scale = 0.2,
  lib.norm = FALSE,
  out.prob = 0.05,
  out.facLoc = 4,
  out.facScale = 0.5,
  group.prob = 1,
  de.prob = 0.1,
  de.downProb = 0.5,
  de.facLoc = 0.1,
  de.facScale = 0.4,
  bcv.common = 0.1,
  bcv.df = 60,
  dropout.type = "none",
  dropout.mid = 0,
  dropout.shape = -1,
  path.from = 0,
  path.nSteps = 100,
  path.skew = 0.5,
  path.nonlinearProb = 0.1,
  path.sigmaFac = 0.8,
  seed = 333,
  ncomp = 10,
  n_hvgs = 500
)
}
\arguments{
\item{nGenes}{Number of genes.}

\item{batchCells}{Number of cells per batch.}

\item{batch.facLoc}{Batch factor location parameter.}

\item{batch.facScale}{Batch factor scale parameter.}

\item{batch.rmEffect}{Remove batch effect.}

\item{mean.rate}{Mean rate.}

\item{mean.shape}{Mean shape.}

\item{lib.loc}{Library size location parameter.}

\item{lib.scale}{Library size scale parameter.}

\item{lib.norm}{Library size distribution.}

\item{out.prob}{Expression outlier probability}

\item{out.facLoc}{Expression outlier factor location}

\item{out.facScale}{Expression outlier factor scale}

\item{group.prob}{Group probabilities}

\item{de.prob}{Differential expression probability}

\item{de.downProb}{Down-regulation probability}

\item{de.facLoc}{DE factor location}

\item{de.facScale}{DE factor scale}

\item{bcv.common}{Common Biological Coefficient of Variation}

\item{bcv.df}{BCV Degrees of Freedom}

\item{dropout.type}{Dropout type}

\item{dropout.mid}{Dropout mid point}

\item{dropout.shape}{Dropout shape}

\item{path.from}{Path origin}

\item{path.nSteps}{Number of steps}

\item{path.skew}{Path skew}

\item{path.nonlinearProb}{Non-linear probability}

\item{path.sigmaFac}{Path skew}

\item{seed}{Random seed}

\item{ncomp}{Number of principal component}

\item{n_hvgs}{Number of higly variable genes.}
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
The function allows to simulated single-cell RNA-seq data using Splatter
package, normalize data, select highly variable genes and
compute Principal Component Analysis.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 50),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)

}
