% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein_distance}
\alias{wasserstein_distance}
\title{Wasserstein distance}
\usage{
wasserstein_distance(input, batch, reduction, rep, mc_cores)
}
\arguments{
\item{input}{A \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{batch}{A string specifying batch variable.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{rep}{Number of times the Wasserstein distance is calculated.}

\item{mc_cores}{The number of cores to use.}
}
\value{
A numeric value.
}
\description{
The Wasserstein distance measures the minimal transport needed to shift one
distribution to match another.
}
\examples{
sim <- simulate_data(
  n_genes = 1000, batch_cells = c(130, 110),
  group_prob = c(0.5, 0.5), n_hvgs = 500,
  compute_pca = TRUE, output_format = "SingleCellExperiment"
)
wass <- wasserstein_distance(
  input = sim, batch = "Batch",
  reduction = "PCA", rep = 1, mc_cores = 1
)

}
