% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein_distance}
\alias{wasserstein_distance}
\title{Wasserstein distance}
\usage{
wasserstein_distance(input, batch, reduction, rep, mc.cores)
}
\arguments{
\item{input}{A \linkS4class{SingleCellExperiment} object.}

\item{batch}{A string specifying batch variable.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{rep}{Number of times the Wasserstein distance is calculated.}

\item{mc.cores}{The number of cores to use.}
}
\value{
A numeric value.
}
\description{
The Wasserstein distance measures the minimal transport needed to shift one
distribution to match another.
}
\examples{
sim <- simulated_data(nGenes = 1000, batchCells = c(150, 50),
                      group.prob = c(0.5, 0.5), n_hvgs = 1000, ncomp = 10)
wass <- wasserstein_distance(input = sim, batch = "Batch",
                             reduction = "PCA", rep = 1, mc.cores = 1)

}
