% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_distance.R
\name{wasserstein_distance}
\alias{wasserstein_distance}
\title{Wasserstein distance}
\usage{
wasserstein_distance(input, batch, reduction, rep, mc.cores)
}
\arguments{
\item{input}{A `SingleCellExperiment` object.}

\item{batch}{A string specifying batch variable.}

\item{reduction}{A string specifying the dimensional reduction.}

\item{rep}{Number of times the Wasserstein distance is calculated.}

\item{mc.cores}{The number of cores to use.}
}
\description{
Wasserstein distance
}
