% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{wasserstein_metric}
\alias{wasserstein_metric}
\title{Title}
\usage{
wasserstein_metric(obj, batch, reduction, n, rep)
}
\arguments{
\item{obj}{a SingleCellExperiment object}

\item{batch}{a string specifying the batch}

\item{reduction}{a string specifying the dimension reduction name}

\item{n}{number of random selection cells}

\item{rep}{number of random selection replicates}
}
\description{
Title
}
